+++ modules/accredited_representative_portal/spec/services/accredited_representative_portal/power_of_attorney_request_service/create/form_data_adapter_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe PowerOfAttorneyRequestService::Create::FormDataAdapter do
+RSpec.describe AccreditedRepresentativePortal::PowerOfAttorneyRequestService::Create::FormDataAdapter do
   describe '#call' do
     subject { described_class.new(data: data, dependent: dependent, service_branch: service_branch) }
